Revert "libsystemd-bus: use assert_return"
This reverts commit f7e2bd5a8070ba86cba6bcbf7d1c9a8173d846d4.

Most of these checks are not programming errors, but happen during
normal runtime. For example bus_kernel_pop_memfd() is called all the
time on non-kdbus systems and is supposed to quickly fail if kdbus is
not available. However, assert_return() makes this failure
expensive, and hence has no place here. With the most recent change to
assert_return() it will even log a debug message, which should never
happen here.