loginctl: correctly show session IDs on session-status
Commit f8f14b3654bcd introduced a regression that makes
loginctl session-status to not show the correct session ID(s)

In print_session_status_info() the map[] array, element "Seat" receives
the offset of the "id" in "SessionStatusInfo" struct instead of the
offset of the "seat" member.

This will cause prop_map_first_of_struct() function to overwrite the
SessionStatusInfo.id memory with seats if there are any.

Fix this typo by using the "seat" member.

Before:
 - tixxdz (1000)
          Since: Sat 2013-12-21 10:07:23 CET; 5h 26min ago
         Leader: 1265 (sshd)

After:
1 - tixxdz (1000)
           Since: Sat 2013-12-21 10:07:23 CET; 5h 26min ago
          Leader: 1265 (sshd)