smack: relabel directories and files created by systemd
Systemd creates directories in /dev. These directories will
get the label of systemd, which is the label of the System
domain, which is not accessable to everyone. Relabel the
directories, files and symlinks created so that they can be
generally used.

Based on a patch by Casey Schaufler <casey@schaufler-ca.com>.