sd-dhcp-client: respect TP_STATUS_CSUMNOTREADY
If an UDP packet has not passed through a hardware device, its checksum may not
have been computed. This is exposed through the TP_STATUS_CSUMNOTREADY sockopt.

When using raw sockets, skip checksum validation when TP_STATUS_CSUMNOTREADY
is set.

This is necessary for dhcp to work directly over a veth tunnel, e.g. as done
in systemd-nspawn.